MY_RESEARCH2/requirements.txt
plotly
matplotlib
seaborn
scikit-learn
google-generativeai  
scipy
streamlit
pandas
numpy
umap-learn
requests
dataclasses; python_version<"3.7"
